[fix] 1차 auth
