+++ src/domain/auth/domain/service/AuthVerify.Service.ts
@@ -43,26 +43,6 @@ export class AuthVerifyService{
             throw new AuthException(AuthErrorCode.IDENTIFIER_IS_INCOREECT);
     }
 
-
-
-    public validateIdentifier(userData: User){
-        if(checkData(userData)){
-            throw new AuthException(AuthErrorCode.INVALIDATE_IDENTIFIER);
-        }
-    }
-
-    public validateEmail(userData: User){
-        if(checkData(userData)){
-            throw new AuthException(AuthErrorCode.INVALIDATE_EMAIL);
-        }
-    }
-
-    public validateNickname(affiliationData: Affiliation){
-        if(checkData(affiliationData)){
-            throw new AuthException(AuthErrorCode.INVALIDATE_NICKNAME);
-        }
-    }
-
     public signVerifyToken(accessTokenVerifyResult: boolean, refreshTokenVerifyesult: boolean){
         this.signVerifyAccessToken(accessTokenVerifyResult);
         this.signVerifyRefreshToken(refreshTokenVerifyesult);
